💥 pathの設定をv2/commonに移動
