+++ src-electron/v2/core/const.ts
@@ -9,15 +9,3 @@ import { Path } from '../util/binary/path';
 export const defaultRuntimeSettings: RuntimeSettings = {
   memory: [2, 'GB'],
 };
-
-const mainPath = (
-  process.env.DEBUGGING || !app
-    ? new Path('userData')
-    : new Path(app?.getPath('userData'))
-).absolute();
-
-const sourcePath = mainPath.child('serverstarter/source');
-
-export const datapackSourcePath = sourcePath.child('datapack');
-
-export const serverSourcePath = sourcePath.child('server');
