+++ src-electron/v2/core/setup.ts
@@ -1,6 +1,6 @@
+import { datapackSourcePath, serverSourcePath } from '../common/paths';
 import { DatapackContainer } from '../source/datapack/datapack';
 import { ServerContainer } from '../source/server/server';
-import { datapackSourcePath, serverSourcePath } from './const';
 
 /**
  * coreが必要とするsourceの初期化を行う
